BE-26, refactor:  merge SwaggerConfigs into one class
